+++ src/main/java/sk/upjs/drpaz/storage/mysql/MysqlProductDao.java
@@ -16,6 +16,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import sk.upjs.drpaz.storage.dao.ProductDao;
 import sk.upjs.drpaz.storage.entities.Category;
 import sk.upjs.drpaz.storage.entities.Product;
+import sk.upjs.drpaz.storage.exceptions.EntityAlreadyReferencedInDatabaseException;
 
 public class MysqlProductDao implements ProductDao {
 
@@ -86,11 +87,15 @@ public class MysqlProductDao implements ProductDao {
 
 	}
 
-	public boolean delete(long id) {
-		jdbcTemplate.update("DELETE FROM purchase_item WHERE product_id = " + id);
-		jdbcTemplate.update("DELETE FROM product_has_category WHERE product_id = " + id);
-		int changed = jdbcTemplate.update("DELETE FROM product WHERE id = " + id);
-		return changed == 1;
+	//TODO need to update UnitTEST
+	public boolean delete(long id) throws EntityAlreadyReferencedInDatabaseException {
+		if(!checkIfCanDelete(id)) {
+			throw new EntityAlreadyReferencedInDatabaseException("Product with id: " + id + " is already referenced in DB.");
+		}else {
+			jdbcTemplate.update("DELETE FROM product_has_category WHERE product_id = " + id);
+			int changed = jdbcTemplate.update("DELETE FROM product WHERE id = " + id);
+			return changed == 1;
+		}
 	}
 
 	@Override
@@ -111,6 +116,20 @@ public class MysqlProductDao implements ProductDao {
 		//TODO test, message me if you dont want to write tests, maybe i can try ):
 		String sql = "SELECT id, name, price, quantity, alert_quantity, description FROM product p JOIN product_has_category phc ON p.id=phc.product_id WHERE category_id= " + category.getId();
 		return jdbcTemplate.query(sql, new ProductRowMapper());
+	}
+	
+	//TODO tests for this
+	//THIS IS FOR CHECKING IF WE CAN DELETE Product
+	@Override
+	public boolean checkIfCanDelete(long id) throws NullPointerException, NoSuchElementException {
+		String sql = "SELECT COUNT(*) FROM purchase_item WHERE product_id = " + id;
+		int count = jdbcTemplate.queryForObject(sql, Integer.class);
+		
+		if (count == 0) {
+			return true;
+		}else {
+			return false;
+		}
 	}
 
 }
